Moved ECommerce Core objects
