+++ GoogleAnalytics/Requests/ECommerce/GAECommerceCartRemoveRequest.php
@@ -4,10 +4,9 @@
 namespace ScAnalytics\GoogleAnalytics\Requests\ECommerce;
 
 
-use ScAnalytics\Analytics;
 use ScAnalytics\Core\AnalyticsConfig;
+use ScAnalytics\Core\ECommerce\Product;
 use ScAnalytics\Core\HelperFunctions;
-use ScAnalytics\Core\Product;
 use ScAnalytics\GoogleAnalytics\GAParameter;
 use ScAnalytics\GoogleAnalytics\Requests\GAEventRequest;
