+++ GoogleAnalytics/Requests/ECommerce/GAECommerceProductClickRequest.php
@@ -5,8 +5,8 @@ namespace ScAnalytics\GoogleAnalytics\Requests\ECommerce;
 
 
 use ScAnalytics\Core\AnalyticsConfig;
+use ScAnalytics\Core\ECommerce\Product;
 use ScAnalytics\Core\HelperFunctions;
-use ScAnalytics\Core\Product;
 use ScAnalytics\GoogleAnalytics\GAParameter;
 use ScAnalytics\GoogleAnalytics\Requests\GAEventRequest;
