+++ Matomo/Matomo.php
@@ -7,10 +7,10 @@ use ScAnalytics\Analytics;
 use ScAnalytics\Core\AnalyticsConfig;
 use ScAnalytics\Core\AnalyticsHandler;
 use ScAnalytics\Core\ARequest;
+use ScAnalytics\Core\ECommerce\Product;
+use ScAnalytics\Core\ECommerce\Transaction;
 use ScAnalytics\Core\HelperFunctions;
 use ScAnalytics\Core\PageData;
-use ScAnalytics\Core\Product;
-use ScAnalytics\Core\Transaction;
 use ScAnalytics\Matomo\Requests\ECommerce\MECommerceCartUpdateRequest;
 use ScAnalytics\Matomo\Requests\ECommerce\MECommerceCheckoutStepRequest;
 use ScAnalytics\Matomo\Requests\ECommerce\MECommerceProductClickRequest;
