+++ Matomo/Requests/ECommerce/MECommerceCheckoutStepRequest.php
@@ -4,8 +4,8 @@
 namespace ScAnalytics\Matomo\Requests\ECommerce;
 
 
+use ScAnalytics\Core\ECommerce\Product;
 use ScAnalytics\Core\PageData;
-use ScAnalytics\Core\Product;
 use ScAnalytics\Matomo\MParameter;
 use ScAnalytics\Matomo\Requests\MPageViewRequest;
