+++ Matomo/Requests/ECommerce/MECommerceRequest.php
@@ -7,8 +7,8 @@ namespace ScAnalytics\Matomo\Requests\ECommerce;
 use JsonException;
 use Money\Money;
 use ScAnalytics\Analytics;
+use ScAnalytics\Core\ECommerce\Product;
 use ScAnalytics\Core\HelperFunctions;
-use ScAnalytics\Core\Product;
 use ScAnalytics\Matomo\MParameter;
 use ScAnalytics\Matomo\Requests\MRequest;
