+++ Tests/GoogleAnalytics/GoogleAnalyticsTest.php
@@ -9,10 +9,10 @@ use ReflectionClass;
 use ReflectionException;
 use ScAnalytics\Analytics;
 use ScAnalytics\Core\AnalyticsConfig;
+use ScAnalytics\Core\ECommerce\Product;
+use ScAnalytics\Core\ECommerce\Transaction;
 use ScAnalytics\Core\PageData;
-use ScAnalytics\Core\Product;
 use ScAnalytics\Core\Scope;
-use ScAnalytics\Core\Transaction;
 use ScAnalytics\GoogleAnalytics\GoogleAnalytics;
 use ScAnalytics\GoogleAnalytics\Requests\ECommerce\GAECommerceCartAddRequest;
 use ScAnalytics\GoogleAnalytics\Requests\ECommerce\GAECommerceCartRemoveRequest;
