+++ Tests/GoogleAnalytics/Requests/ECommerce/GAECommercePurchaseRequestTest.php
@@ -4,16 +4,15 @@ namespace GoogleAnalytics\Requests\ECommerce;
 
 use Money\Currency;
 use Money\Money;
+use PHPUnit\Framework\TestCase;
 use ReflectionClass;
 use ReflectionException;
 use ScAnalytics\Analytics;
-use ScAnalytics\Core\Product;
+use ScAnalytics\Core\ECommerce\Product;
+use ScAnalytics\Core\ECommerce\Transaction;
 use ScAnalytics\Core\Scope;
-use ScAnalytics\Core\Transaction;
 use ScAnalytics\GoogleAnalytics\GAParameter;
-use ScAnalytics\GoogleAnalytics\Requests\ECommerce\GAECommerceCartRemoveRequest;
 use ScAnalytics\GoogleAnalytics\Requests\ECommerce\GAECommercePurchaseRequest;
-use PHPUnit\Framework\TestCase;
 
 /**
  * Tests the GAECommercePurchaseRequest class.
