+++ Tests/GoogleAnalytics/Requests/ECommerce/GAECommerceRefundRequestTest.php
@@ -4,15 +4,15 @@ namespace GoogleAnalytics\Requests\ECommerce;
 
 use Money\Currency;
 use Money\Money;
+use PHPUnit\Framework\TestCase;
 use ReflectionClass;
 use ReflectionException;
 use ScAnalytics\Analytics;
-use ScAnalytics\Core\Product;
+use ScAnalytics\Core\ECommerce\Product;
+use ScAnalytics\Core\ECommerce\Transaction;
 use ScAnalytics\Core\Scope;
-use ScAnalytics\Core\Transaction;
 use ScAnalytics\GoogleAnalytics\GAParameter;
 use ScAnalytics\GoogleAnalytics\Requests\ECommerce\GAECommerceRefundRequest;
-use PHPUnit\Framework\TestCase;
 
 /**
  * Tests the GAECommerceRefundRequest class.
