+++ Tests/Matomo/MatomoTest.php
@@ -9,10 +9,10 @@ use ReflectionClass;
 use ReflectionException;
 use ScAnalytics\Analytics;
 use ScAnalytics\Core\AnalyticsConfig;
+use ScAnalytics\Core\ECommerce\Product;
+use ScAnalytics\Core\ECommerce\Transaction;
 use ScAnalytics\Core\PageData;
-use ScAnalytics\Core\Product;
 use ScAnalytics\Core\Scope;
-use ScAnalytics\Core\Transaction;
 use ScAnalytics\Matomo\Matomo;
 use ScAnalytics\Matomo\Requests\ECommerce\MECommerceCartUpdateRequest;
 use ScAnalytics\Matomo\Requests\ECommerce\MECommerceCheckoutStepRequest;
