+++ Tests/Matomo/Requests/ECommerce/MECommercePurchaseRequestTest.php
@@ -4,15 +4,15 @@ namespace Matomo\Requests\ECommerce;
 
 use Money\Currency;
 use Money\Money;
+use PHPUnit\Framework\TestCase;
 use ReflectionClass;
 use ReflectionException;
 use ScAnalytics\Analytics;
-use ScAnalytics\Core\Product;
+use ScAnalytics\Core\ECommerce\Product;
+use ScAnalytics\Core\ECommerce\Transaction;
 use ScAnalytics\Core\Scope;
-use ScAnalytics\Core\Transaction;
 use ScAnalytics\Matomo\MParameter;
 use ScAnalytics\Matomo\Requests\ECommerce\MECommercePurchaseRequest;
-use PHPUnit\Framework\TestCase;
 
 /**
  * Tests the MECommercePurchaseRequest class.
